Packages: move OffSetters to .cell and FrameAwareCutoutImageCell to top-level
