+++ indicator/src/main/java/com/fuzz/indicator/ImageCellGenerator.java
@@ -22,7 +22,6 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 
 import com.fuzz.indicator.cell.CutoutCell;
-import com.fuzz.indicator.clip.FrameAwareCutoutImageCell;
 
 /**
  * Default implementation of {@link CutoutCellGenerator}.
