+++ indicator/src/main/java/com/fuzz/indicator/cell/CutoutCell.java
@@ -20,7 +20,6 @@ import android.view.View;
 
 import com.fuzz.indicator.CutoutCellGenerator;
 import com.fuzz.indicator.CutoutViewIndicator;
-import com.fuzz.indicator.OffSetters;
 import com.fuzz.indicator.proxy.IndicatorOffsetEvent;
 
 /**
