+++ indicator/src/main/java/com/fuzz/indicator/cell/CutoutTextCell.java
@@ -21,8 +21,6 @@ import android.text.SpannableString;
 import android.text.Spanned;
 import android.widget.TextView;
 
-import com.fuzz.indicator.cell.TypicalCutoutCell;
-import com.fuzz.indicator.OffSetters;
 import com.fuzz.indicator.proxy.IndicatorOffsetEvent;
 
 /**
