+++ indicator/src/main/java/com/fuzz/indicator/cell/OffSetters.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fuzz.indicator;
+package com.fuzz.indicator.cell;
 
 import android.graphics.Matrix;
 import android.os.Build;
